Watching tests now works, and we get working test viewer!
